safe_scheduled_autoscaling/app/requirements.txt
Flask==3.1.0
google-auth==2.38.0
requests==2.32.3
gunicorn==23.0.0
